Move httprule back into the public API
Fixes #1168
