+++ protoc-gen-grpc-gateway/httprule/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
         "parse.go",
         "types.go",
     ],
-    importpath = "github.com/grpc-ecosystem/grpc-gateway/internal/httprule",
+    importpath = "github.com/grpc-ecosystem/grpc-gateway/protoc-gen-grpc-gateway/httprule",
     deps = [
         "//utilities:go_default_library",
         "@com_github_golang_glog//:go_default_library",
