+++ protoc-gen-grpc-gateway/internal/gengateway/BUILD.bazel
@@ -30,8 +30,8 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//internal/httprule:go_default_library",
         "//protoc-gen-grpc-gateway/descriptor:go_default_library",
+        "//protoc-gen-grpc-gateway/httprule:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@io_bazel_rules_go//proto/wkt:descriptor_go_proto",
     ],
