+++ protoc-gen-swagger/genswagger/BUILD.bazel
@@ -36,8 +36,8 @@ go_test(
     srcs = ["template_test.go"],
     embed = [":go_default_library"],
     deps = [
-        "//internal/httprule:go_default_library",
         "//protoc-gen-grpc-gateway/descriptor:go_default_library",
+        "//protoc-gen-grpc-gateway/httprule:go_default_library",
         "//protoc-gen-swagger/options:go_default_library",
         "@com_github_golang_protobuf//proto:go_default_library",
         "@io_bazel_rules_go//proto/wkt:any_go_proto",
